Extract InfinispanServerRule into a separate reusable InfinispanServerUtil class.
