+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/AbstractClusteringTestCase.java
@@ -21,20 +21,13 @@
  */
 package org.jboss.as.test.clustering.cluster;
 
-import java.net.URISyntaxException;
-import java.nio.file.Paths;
 import java.util.AbstractMap;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
 import org.apache.http.Header;
 import org.apache.http.HttpResponse;
-import org.infinispan.server.test.core.ServerRunMode;
-import org.infinispan.server.test.core.TestSystemPropertyNames;
-import org.infinispan.server.test.junit4.InfinispanServerRule;
-import org.infinispan.server.test.junit4.InfinispanServerRuleBuilder;
 import org.jboss.arquillian.container.spi.Container;
 import org.jboss.arquillian.container.spi.ContainerRegistry;
 import org.jboss.arquillian.container.test.api.Deployer;
@@ -45,7 +38,6 @@ import org.jboss.as.test.shared.TimeoutUtil;
 import org.jboss.logging.Logger;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.rules.TestRule;
 
 /**
  * Base implementation for every clustering test which guarantees a framework contract as follows:
@@ -110,30 +102,6 @@ public abstract class AbstractClusteringTestCase {
     public static final int INFINISPAN_SERVER_PORT = 11222;
     public static final String INFINISPAN_APPLICATION_USER = "testsuite-application-user";
     public static final String INFINISPAN_APPLICATION_PASSWORD = "testsuite-application-password";
-    public static final InfinispanServerRule INFINISPAN_SERVER_RULE;
-
-    static {
-        String profile = (INFINISPAN_SERVER_PROFILE == null || INFINISPAN_SERVER_PROFILE.isEmpty()) ? INFINISPAN_SERVER_PROFILE_DEFAULT : INFINISPAN_SERVER_PROFILE;
-        // Workaround for "ISPN-13107 ServerRunMode.FORKED yields InvalidPathException with relative server config paths on Windows platform" by using absolute file path which won't get mangled.
-        String absoluteConfigurationFile = null;
-        try {
-            absoluteConfigurationFile = Paths.get(Objects.requireNonNull(AbstractClusteringTestCase.class.getClassLoader().getResource(profile)).toURI()).toFile().toString();
-        } catch (URISyntaxException ignore) {
-        }
-
-        INFINISPAN_SERVER_RULE = InfinispanServerRuleBuilder
-                .config(absoluteConfigurationFile)
-                .property(TestSystemPropertyNames.INFINISPAN_TEST_SERVER_DIR, INFINISPAN_SERVER_HOME)
-                .property("infinispan.client.rest.auth_username", "testsuite-driver-user")
-                .property("infinispan.client.rest.auth_password", "testsuite-driver-password")
-                .numServers(1)
-                .runMode(ServerRunMode.FORKED)
-                .build();
-    }
-
-    public static TestRule infinispanServerTestRule() {
-        return INFINISPAN_SERVER_RULE;
-    }
 
     // Undertow-based WildFly load-balancer
     public static final String LOAD_BALANCER_1 = "load-balancer-1";
