+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/web/remote/AbstractHotRodPersistenceWebFailoverTestCase.java
@@ -25,6 +25,7 @@ package org.jboss.as.test.clustering.cluster.web.remote;
 import org.infinispan.configuration.cache.CacheMode;
 import org.infinispan.transaction.TransactionMode;
 import org.jboss.as.test.clustering.ClusterTestUtil;
+import org.jboss.as.test.clustering.InfinispanServerUtil;
 import org.jboss.as.test.clustering.cluster.web.AbstractWebFailoverTestCase;
 import org.jboss.as.test.clustering.single.web.Mutable;
 import org.jboss.as.test.clustering.single.web.SimpleServlet;
@@ -44,7 +45,7 @@ import org.junit.rules.TestRule;
 public abstract class AbstractHotRodPersistenceWebFailoverTestCase extends AbstractWebFailoverTestCase {
 
     @ClassRule
-    public static final TestRule INFINISPAN_SERVER_RULE = infinispanServerTestRule();
+    public static final TestRule INFINISPAN_SERVER_RULE = InfinispanServerUtil.infinispanServerTestRule();
 
     static Archive<?> getDeployment(String deploymentName, String deploymentDescriptor) {
         WebArchive war = ShrinkWrap.create(WebArchive.class, deploymentName);
