+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/web/remote/AbstractHotRodSessionActivationTestCase.java
@@ -22,6 +22,7 @@
 
 package org.jboss.as.test.clustering.cluster.web.remote;
 
+import org.jboss.as.test.clustering.InfinispanServerUtil;
 import org.jboss.as.test.clustering.cluster.web.AbstractSessionActivationTestCase;
 import org.junit.ClassRule;
 import org.junit.rules.TestRule;
@@ -33,7 +34,7 @@ import org.junit.rules.TestRule;
 public abstract class AbstractHotRodSessionActivationTestCase extends AbstractSessionActivationTestCase {
 
     @ClassRule
-    public static final TestRule INFINISPAN_SERVER_RULE = infinispanServerTestRule();
+    public static final TestRule INFINISPAN_SERVER_RULE = InfinispanServerUtil.infinispanServerTestRule();
 
     protected AbstractHotRodSessionActivationTestCase(boolean transactional) {
         super(transactional);
