+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/web/remote/AbstractHotRodSessionExpirationTestCase.java
@@ -23,6 +23,7 @@
 package org.jboss.as.test.clustering.cluster.web.remote;
 
 import org.infinispan.transaction.TransactionMode;
+import org.jboss.as.test.clustering.InfinispanServerUtil;
 import org.jboss.as.test.clustering.cluster.web.expiration.SessionExpirationTestCase;
 import org.junit.ClassRule;
 import org.junit.rules.TestRule;
@@ -33,7 +34,7 @@ import org.junit.rules.TestRule;
 public abstract class AbstractHotRodSessionExpirationTestCase extends SessionExpirationTestCase {
 
     @ClassRule
-    public static final TestRule INFINISPAN_SERVER_RULE = infinispanServerTestRule();
+    public static final TestRule INFINISPAN_SERVER_RULE = InfinispanServerUtil.infinispanServerTestRule();
 
     public AbstractHotRodSessionExpirationTestCase() {
         super(TransactionMode.NON_TRANSACTIONAL);
